clear AppDatabase & dbModule
